# bandersnatch/fp: more performant Inv implementation